fix tests and review issues
